feat: 实现AI训练专用接口 (RL Interface)
实现两个关键的AI训练接口：

✅ 1. to_tensor() 方法
   - 位置：PyGameState::to_tensor()
   - 功能：将 GameState 扁平化为 Vec<f32>
   - 包含：
     * 4D tensor 扁平化（64 × 4 × 9 = 2304 个浮点数）
     * 牌池可见性特征（108 个浮点数，每张牌的出现次数 0-4）
   - 总计：2412 个浮点数
   - 牌池可见性统计：
     * 所有玩家的手牌
     * 已碰/杠的牌（碰：3 张，杠：4 张）
     * 弃牌历史

✅ 2. get_action_mask() 方法
   - 位置：PyActionMask::get_action_mask()
   - 功能：返回动作掩码数组
   - 返回：Vec<f32>，长度为 434
   - 转换：bool → f32（true → 1.0，false → 0.0）
   - 动作空间：
     * 索引 0-107: 出牌动作
     * 索引 108-215: 碰牌动作
     * 索引 216-323: 杠牌动作
     * 索引 324-431: 胡牌动作
     * 索引 432: 摸牌动作
     * 索引 433: 过动作

修改文件：
- rust/src/python/game_state.rs: 添加 to_tensor() 和 compute_tile_visibility() 方法
- rust/src/python/action_mask.rs: 添加 get_action_mask() 静态方法
- rust/src/python/tensor.rs: 修复索引类型问题（u8 → usize）
- rust/RL_INTERFACE_AUDIT.md: 审计文档，详细说明实现细节

测试结果：
- 所有测试通过（72 个测试）
- 代码已通过编译检查

+++ rust/src/python/tensor.rs
@@ -58,7 +58,7 @@ pub fn state_to_tensor(
             for (tile, &tile_count) in player.hand.tiles_map() {
                 if tile_count == count {
                     let (suit_idx, rank_idx) = tile_to_indices(tile);
-                    data[[count - 1, suit_idx, rank_idx]] = 1.0;
+                    data[[(count - 1) as usize, suit_idx, rank_idx]] = 1.0;
                 }
             }
         }
